CL:
* Changed constructors. number of threads fixed to 1 and 3.
* Moved threadEventDispatcher test
* Fix compilation
