+++ src/engine/source/base/CMakeLists.txt
@@ -27,6 +27,7 @@ target_link_libraries(base
     PUBLIC
     fmt::fmt-header-only
     spdlog::spdlog
+    RocksDB::rocksdb
 )
 
 # Tests
@@ -50,6 +51,7 @@ add_executable(base_utest
     ${UNIT_SRC_DIR}/utils/wazuhProtocol/wazuhRequest_test.cpp
     ${UNIT_SRC_DIR}/utils/wazuhProtocol/wazuhResponse_test.cpp
     ${UNIT_SRC_DIR}/utils/timeUtils_test.cpp
+    ${UNIT_SRC_DIR}/utils/threadEventDispatcher_test.cpp
     ${UNIT_SRC_DIR}/parseEvent_test.cpp
     ${UNIT_SRC_DIR}/dotPath_test.cpp
     ${UNIT_SRC_DIR}/json_test.cpp
